Added cast2_layer.cpp
